+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - remove initial / from page slug to use it in the API fetch
+- Blacklist the refresh token on the frontend side (#2265)
 
 ## [4.1.0] - 2023-05-23